requirements.txt
torch>=2.1
transformers>=4.40
datasets>=2.16
numpy>=1.24
scipy>=1.11
scikit-learn>=1.3
pandas>=2.0
pyyaml>=6.0
tqdm
joblib
sae-lens>=3.0
matplotlib>=3.7
seaborn>=0.13
